Windows compatible Makefile

+++ lib/src/alloc.c
@@ -4,7 +4,7 @@
 static void *ts_malloc_default(size_t size) {
   void *result = malloc(size);
   if (size > 0 && !result) {
-    fprintf(stderr, "tree-sitter failed to allocate %zu bytes", size);
+    fprintf(stderr, "tree-sitter failed to allocate %lu bytes", (unsigned long)size);
     exit(1);
   }
   return result;
@@ -13,7 +13,7 @@ static void *ts_malloc_default(size_t size) {
 static void *ts_calloc_default(size_t count, size_t size) {
   void *result = calloc(count, size);
   if (count > 0 && !result) {
-    fprintf(stderr, "tree-sitter failed to allocate %zu bytes", count * size);
+    fprintf(stderr, "tree-sitter failed to allocate %lu bytes", (unsigned long)(count * size));
     exit(1);
   }
   return result;
@@ -22,7 +22,7 @@ static void *ts_calloc_default(size_t count, size_t size) {
 static void *ts_realloc_default(void *buffer, size_t size) {
   void *result = realloc(buffer, size);
   if (size > 0 && !result) {
-    fprintf(stderr, "tree-sitter failed to reallocate %zu bytes", size);
+    fprintf(stderr, "tree-sitter failed to reallocate %lu bytes", (unsigned long)size);
     exit(1);
   }
   return result;
